xds: Move JwtTokenFileCallCredentials to xds.client package and make package-private
Move `JwtTokenFileCallCredentials` from `io.grpc.auth` to `io.grpc.xds.client` and make it package-private as it is only used in xds.

To support this move without breaking tests:
- Moved `JwtTokenFileCallCredentialsTest` to the same package (`io.grpc.xds.client`).
- Created `BootstrapperImplTest` in `io.grpc.xds.client` to test the base `BootstrapperImpl` parsing logic.
- Moved `call_creds` parsing tests (which reference `JwtTokenFileCallCredentials`) from `GrpcBootstrapperImplTest` to the new `BootstrapperImplTest`.
