+++ xds/src/test/java/io/grpc/xds/client/JwtTokenFileCallCredentialsTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.auth;
+package io.grpc.xds.client;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
